sentimentsync/requirements.txt
streamlit
pandas
scikit-learn
scipy
plotly
seaborn
matplotlib
google-generativeai
typing-extensions